requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
boto3==1.43.0
botocore==1.43.0
certifi==2026.4.22
cffi==2.0.0
click==8.3.3
colorama==0.4.6
cryptography==47.0.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.136.1
fastapi-cli==0.0.24
fastapi-cloud-cli==0.17.1
fastar==0.11.0
greenlet==3.5.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.13
jinja2==3.1.6
jmespath==1.1.0
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
minio==7.2.20
psycopg2-binary==2.9.12
pwdlib==0.3.0
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.13.3
pydantic-core==2.46.3
pydantic-extra-types==2.11.1
pydantic-settings==2.14.0
pygments==2.20.0
pyjwt==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.27
pyyaml==6.0.3
rich==15.0.0
rich-toolkit==0.19.7
rignore==0.7.6
s3transfer==0.17.0
sentry-sdk==2.58.0
shellingham==1.5.4
six==1.17.0
sqlalchemy==2.0.49
starlette==1.0.0
typer==0.25.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.46.0
watchfiles==1.1.1
websockets==16.0
